socialmedia/serializer.py
from rest_framework import serializers
from .models import User,FriendRequest
from django.core.validators import EmailValidator
from rest_framework_simplejwt.serializers import TokenRefreshSerializer
from rest_framework_simplejwt.state import token_backend


class UserSerializer(serializers.ModelSerializer):
    
    class Meta:
        model = User
        fields = "__all__"


class FriendRequestSerializer(serializers.ModelSerializer):
    class Meta:
        model = FriendRequest
        fields = ['id', 'from_user', 'to_user', 'status']

class UserRegisterationSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['email', 'username', 'password']
        extra_kwargs = {
            'password': {'write_only': True}
        }

class UserLoginSerializer(serializers.ModelSerializer):
    class Meta:
      model = User
      fields = ['email', 'username']

class CustomTokenRefreshSerializer(TokenRefreshSerializer):
    def validate(self, attrs):
        # The default result (access/refresh tokens)
        data = super(CustomTokenRefreshSerializer, self).validate(attrs)
        decoded_payload = token_backend.decode(data['access'], verify=True)
        response_data = {
            'message': "New access token generated",
            'data' : data
        }
        return response_data
