requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
bcrypt==4.0.1
decorator==5.1.1
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
executing==2.0.1
ipdb==0.13.13
ipython==8.17.2
jedi==0.19.1
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
wcwidth==0.2.9
django-environ==0.11.2
django-cors-headers==4.3.0
environ
django-treebeard
django-storages==1.14.2
django-filter==23.3